Fix the setup script to look for db existance before deleting it

+++ setupCardBase.sh
@@ -1,2 +1,5 @@
-rm cards.db
+if [[ -f 'cards.db' ]] ; then
+	rm cards.db
+fi
 ruby cardsdb.rb
+
